Fix minpoint bug
Passing a value to the --minpoints option it is saved as a string. This
caused the option to fail in the makeouttree function where the string
had to be added to an int. The default value works because it is defined
internally as an int.

Just added a line to explicitly define minpoints as an int.